refactored RestGetSuggestionsAction such that request parsing happens in GetSuggestionsRequest instead
