+++ alerting/src/main/kotlin/org/opensearch/alerting/action/GetSuggestionsRequest.kt
@@ -13,6 +13,8 @@ import org.opensearch.alerting.model.suggestions.suggestioninputs.util.Suggestio
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
 import org.opensearch.common.xcontent.XContentParser
+import org.opensearch.common.xcontent.XContentParserUtils
+import org.opensearch.rest.RestRequest
 import java.io.IOException
 
 class GetSuggestionsRequest : ActionRequest {
@@ -45,5 +47,68 @@ class GetSuggestionsRequest : ActionRequest {
 
     override fun validate(): ActionRequestValidationException? {
         return null
+    }
+
+    companion object {
+        @JvmStatic
+        fun parse(request: RestRequest): GetSuggestionsRequest {
+            var inputType: SuggestionInputType? = null
+            var component: String? = null
+            var hasInput = false
+
+            val xcp = request.contentParser()
+
+            XContentParserUtils.ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.nextToken(), xcp)
+            while (xcp.nextToken() != XContentParser.Token.END_OBJECT) {
+                val fieldName = xcp.currentName()
+                xcp.nextToken()
+
+                when (fieldName) {
+                    SuggestionInput.INPUT_TYPE_FIELD -> {
+                        val rawInputType = xcp.text()
+                        val allowedInputTypes = SuggestionInputType.values().map { it.value }
+                        if (!allowedInputTypes.contains(rawInputType)) {
+                            throw IllegalArgumentException("invalid inputType, must be one of $allowedInputTypes")
+                        }
+                        try {
+                            inputType = SuggestionInputType.enumFromStr(rawInputType)
+                        } catch (e: Exception) {
+                            throw IllegalArgumentException("invalid inputType, must be one of $allowedInputTypes")
+                        }
+                    }
+                    SuggestionInput.COMPONENT_FIELD -> component = xcp.text()
+                    SuggestionInput.INPUT_FIELD -> {
+                        XContentParserUtils.ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.currentToken(), xcp)
+                        hasInput = true
+                        xcp.skipChildren()
+                    }
+                    else -> throw IllegalArgumentException("request body must contain only input, inputType, and component fields")
+                }
+            }
+
+            if (inputType == null || component == null || !hasInput) {
+                throw IllegalArgumentException("request body must contain input, inputType, and component fields")
+            }
+
+            val newXcp = request.contentParser()
+            prepareXcp(newXcp)
+
+            return GetSuggestionsRequest(inputType, component, newXcp) // newXcp already pointing to beginning of input{} object
+        }
+
+        // prepare by making it point to the start of the "input{}" object rather
+        // than the start of the entire request body
+        private fun prepareXcp(xcp: XContentParser) {
+            XContentParserUtils.ensureExpectedToken(XContentParser.Token.START_OBJECT, xcp.nextToken(), xcp)
+            while (xcp.nextToken() != XContentParser.Token.END_OBJECT) {
+                val fieldName = xcp.currentName()
+                xcp.nextToken()
+                if (fieldName == SuggestionInput.INPUT_FIELD) {
+                    break
+                } else {
+                    xcp.skipChildren()
+                }
+            }
+        }
     }
 }
